+++ cmd/qlueless-assembly-line-api.go
@@ -7,6 +7,8 @@ import (
 	"log"
 	"net/http"
 
+	"github.com/gorilla/mux"
+
 	bat "github.com/PaulioRandall/qlueless-assembly-line-api/internal/app/batches"
 	dict "github.com/PaulioRandall/qlueless-assembly-line-api/internal/app/dictionaries"
 	oai "github.com/PaulioRandall/qlueless-assembly-line-api/internal/app/openapi"
@@ -16,10 +18,16 @@ import (
 // Main is the entry point for the web server
 func main() {
 	log.Println("[Qlueless Assembly Line API]: Starting application")
-	http.HandleFunc("/openapi", oai.OpenAPI_handler)
-	http.HandleFunc("/dictionaries", dict.All_dictionaries_handler)
-	http.HandleFunc("/orders", ord.All_orders_handler)
-	http.HandleFunc("/batches", bat.All_batches_handler)
+
+	gorilla := mux.NewRouter()
+
+	gorilla.HandleFunc("/openapi", oai.OpenAPI_handler)
+	gorilla.HandleFunc("/dictionaries", dict.All_dictionaries_handler)
+	gorilla.HandleFunc("/orders", ord.All_orders_handler)
+	gorilla.HandleFunc("/orders/{order_id}", ord.Single_order_handler)
+	gorilla.HandleFunc("/batches", bat.All_batches_handler)
+
+	http.Handle("/", gorilla)
 
 	log.Println("[Qlueless Assembly Line API]: Starting server")
 	log.Fatal(http.ListenAndServe(":8080", nil))